rag_groq_inference/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
lancedb
wikipedia
beautifulsoup4
pypdf
chromadb
faiss-cpu
arxiv
langchainhub
langchain-groq
# langchain-text-splitters
